+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.11.0] - 2021-11-20
+## [0.12.1] - 2021-11-20
 ### Added
  - base64 command.
  - cp command.
+ - nl command.
+ - -n option for cat command
 ### Changed
  - Reduce mimixbox binary size by compile option (7.5MB --> 5.4MB)
+ - cat/tac command to receive data from standard input when the argument is "-".
 ## [0.9.1] - 2021-11-19
 ### Added
  - basename command.